requirements.txt
loguru
pandas
numpy
scrapy
python-telegram-bot
bertopic
faiss-cpu
sentence-transformers
hdbscan
umap-learn
datasets
scikit-learn
matplotlib
nltk
tabulate
pytest
mypy
black
selenium
tqdm
transformers[torch]
tensorboard
